+++ src/ResourceType.ts
@@ -1,0 +1,1 @@
+export type ResourceType = 'moon' | 'star' | 'atom';