+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Basic card layout
+
 ## [0.0.4] - 2020-05-02
 ### Added
 - HP